4662 - Allow style for table tags in the editor
